Bug fixes
* Fixed metrics
* Finished slice.h
* Decided to continue using Eigen since the operations are vectorized
* Strategy actually works now

+++ cash.cpp
@@ -1,13 +1,15 @@
 #include "cash.h"
 #include "direction.h"
 
+#include <iostream>
+
 CASH::CASH() {}
 
 CASH::CASH(double starting_quantity) {
     quantity = starting_quantity;
 };
 
-double CASH::update_value(double price)
+double CASH::update_value(long price)
 {
     return quantity;
 };
@@ -16,5 +18,7 @@ void CASH::update_position(double order_price, double order_quantity, Direction 
 {
     // self.quantity -= commission;
     // self.total_commission += commission  
+    // std::cout << "You have " << quantity << std::endl;
     quantity = quantity + (-1 * direction * order_quantity * order_price);
+    // std::cout << "Deducted " << (-1 * direction * order_quantity * order_price) << " from the bank." << std::endl;;
 }
