+++ cash.h
@@ -9,7 +9,7 @@ class CASH {
       CASH();
       CASH(double starting_quantity);
       void update_position(double order_price, double order_quantity, Direction direction);
-      double update_value(double price);
+      double update_value(long price);
 };
 
 #endif
