+++ position.cpp
@@ -20,7 +20,7 @@ void Position::update_position(double order_price, double order_quantity, Direct
     trades.push_front(Trade(0, order_price, quantity, direction));
     quantity = quantity + ( direction * order_quantity );
 }
-double Position::update_value(double price)
+double Position::update_value(long price)
 {
     return quantity * price;
 };
