+++ position.h
@@ -17,7 +17,7 @@ class Position {
 
     void update_position(double order_price, double order_quantity, Direction direction);
 
-    double update_value(double price);
+    double update_value(long price);
 
 };
